refactor: migrate LikedPage to React Query with useLikedArtQuery hook
